+++ Portfolio/GameEngineContents/PreCompile.h
@@ -1,19 +1,19 @@
 #pragma once
 
-#define GAMEPAD_DPAD_UP					 0x0001
-#define GAMEPAD_DPAD_DOWN				 0x0002
-#define GAMEPAD_DPAD_LEFT				 0x0004
-#define GAMEPAD_DPAD_RIGHT				 0x0008
-#define GAMEPAD_START					 0x0010
-#define GAMEPAD_BACK					 0x0020
-#define GAMEPAD_LEFT_THUMB				 0x0040
-#define GAMEPAD_RIGHT_THUMB				 0x0080
-#define GAMEPAD_LEFT_SHOULDER			 0x0100
-#define GAMEPAD_RIGHT_SHOULDER			 0x0200
-#define GAMEPAD_A						 0x1000
-#define GAMEPAD_B						 0x2000
-#define GAMEPAD_X						 0x4000
-#define GAMEPAD_Y                        0x8000
+#define GAMEPAD_DPAD_UP        0x0001
+#define GAMEPAD_DPAD_DOWN      0x0002
+#define GAMEPAD_DPAD_LEFT      0x0004
+#define GAMEPAD_DPAD_RIGHT     0x0008
+#define GAMEPAD_START          0x0010
+#define GAMEPAD_BACK           0x0020
+#define GAMEPAD_LEFT_THUMB     0x0040
+#define GAMEPAD_RIGHT_THUMB    0x0080
+#define GAMEPAD_LEFT_SHOULDER  0x0100
+#define GAMEPAD_RIGHT_SHOULDER 0x0200
+#define GAMEPAD_A              0x1000
+#define GAMEPAD_B              0x2000
+#define GAMEPAD_X              0x4000
+#define GAMEPAD_Y              0x8000
 
 
 // std and direct
@@ -55,7 +55,7 @@
 #include <GameEngineCore/GameEngineCollision.h>
 #include <GameEngineCore/GameEngineFontRenderer.h>
 
-//Contents
+// Contents
 #include <GameEngineContents/StageActor.h>
 #include "GravityComponent.h"
 #include "CutScenePlayer.h"
@@ -63,7 +63,6 @@
 #include "LoadingActor.h"
 
 #include "Penitent.h"
-
 
 
 namespace
@@ -78,15 +77,15 @@ namespace
             dir.Move(*(path.begin() + i));
         }
     }
-}
+}  // namespace
 
 struct MetaData
 {
     int   Index;
-    int   PosX;
-    int   PosY;
-    int   Width;
-    int   Height;
+    float PosX;
+    float PosY;
+    float Width;
+    float Height;
     float PivotX;
     float PivotY;
 };
